Fix updateFlag() corrupting content of the register
Using updateFlag() on the flag register can create bad values
for the content of the flag register. Using updateFlag() to set
the zero flag causes the content of the register to be bitwise or'ed
with -128 due to bitmasks being stored as byte. In order to or the
char with the byte, Java first casts the byte into a char, and then
bitwise or the two chars.

-128 converted to char is 0xFF80, and when 0xFF80 is or'ed with the
content of the register, it corrupts the register.

+++ GameDroid/app/src/main/java/creativename/gamedroid/core/CPU.java
@@ -608,13 +608,13 @@ public class CPU {
             SUBTRACTION(0x40),
             ZERO(0x80);
 
-            private final byte bitmask;
+            private final int bitmask;
 
             Flag(int mask) {
-                bitmask = (byte) mask;
+                bitmask = mask;
             }
 
-            public byte getBitmask() {
+            public int getBitmask() {
                 return bitmask;
             }
         }
